requirements.txt
Flask==3.1.2
flask_sqlalchemy==3.1.1
python-dotenv==1.2.1
SQLAlchemy==2.0.45
psycopg2==2.9.11
gunicorn==23.0.0